A5 Process work Day 1
This was my process work for Day 1 of my 2d Platformer. I was able to load a black square png as the player, and use the same black square as the ground with some stretching. I added colliders to both the player and the ground and added rigidbody 2d to the player along with a player movement script for horizontal (x axis) movement and jumping The script uses the update function which allows us to check for inputs on every frame.